detect kernel version
