--- daemon/graphdriver/overlay2/mount.go
@@ -1,14 +1,0 @@
-// +build linux
-
-package overlay2
-
-import "syscall"
-
-// XXX: copied from AUFS
-// Unmount the target specified.
-func Unmount(target string) error {
-	if err := syscall.Unmount(target, 0); err != nil {
-		return err
-	}
-	return nil
-}
